+++ src/instrument/bias_image.py
@@ -10,7 +10,7 @@ def generate_bias_image(channel: SpectroscopyChannel):
     read_noise = channel.read_noise
     ccd_gain = channel.ccd_gain
 
-    bias = (bias_offset + np.random.normal(0.0, read_noise, size=(ny, nx)))
+    bias = (bias_offset + np.random.normal(0.0, read_noise, size=(ny, nx))).astype(np.float32)
     
     logging.info("BIAS IMAGE %s (%d x %d): bias_offset=%g read_noise=%g ccd_gain=%g -> mean=%g std=%g min=%g max=%g", channel.channel_name, nx, ny, bias_offset, read_noise, ccd_gain, bias.mean(), bias.std(), bias.min(), bias.max())
     return bias
